Fixing up remaining visible warnings in xcode (all in test programs).
Expanding or removing some '@@@@' marks.

Copied from Perforce
 Change: 179574
 ServerID: perforce.ravenbrook.com

+++ mps/code/sacss.c
@@ -80,7 +80,7 @@ static mps_res_t stress(mps_class_t class,
   for (k = 0; k < testLOOPS; ++k) {
     /* shuffle all the objects */
     for (i=0; i<testSetSIZE; ++i) {
-      int j = rnd()%(testSetSIZE-i);
+      int j = (int)(rnd()%(unsigned)(testSetSIZE-i));
       void *tp;
       size_t ts;
